+++ library/LibUtilities/LinearAlgebra/NekLinSysIter.h
@@ -77,19 +77,15 @@ public:
 
     LIB_UTILITIES_EXPORT void SetUniversalUniqueMap(
         const Array<OneD, const int> &map);
-    LIB_UTILITIES_EXPORT void setRhsMagnitude(const NekDouble mag)
+
+    LIB_UTILITIES_EXPORT void SetNekLinSysTolerance(const NekDouble in)
     {
-        m_rhs_magnitude = mag;
+        m_NekLinSysTolerance = in;
     }
 
     LIB_UTILITIES_EXPORT void SetNekLinSysMaxIterations(const unsigned int in)
     {
-        m_maxiter = in;
-    }
-
-    LIB_UTILITIES_EXPORT void SetLinSysMaxStorage(const unsigned int in)
-    {
-        m_LinSysMaxStorage = in;
+        m_NekLinSysMaxIterations = in;
     }
 
     LIB_UTILITIES_EXPORT bool IsLocal()
@@ -101,11 +97,11 @@ protected:
     /// Global to universal unique map
     Array<OneD, int> m_map;
 
-    /// Dot product of rhs to normalise stopping criterion
-    NekDouble m_rhs_magnitude = NekConstants::kNekUnsetDouble;
-
-    int m_totalIterations   = 0;
     NekDouble m_prec_factor = 1.0;
+
+    NekDouble m_NekLinSysTolerance;
+    int m_NekLinSysMaxIterations;
+    int m_totalIterations = 0;
 
     // This is the maximum number of solution vectors that can be stored
     // For example, in gmres, it is the max number of Krylov space
@@ -116,10 +112,12 @@ protected:
     // Boolean to identify if iteration acts on local storage
     bool m_isLocal;
 
+    void v_InitObject() override;
+
     void Set_Rhs_Magnitude(const Array<OneD, NekDouble> &pIn);
     void SetUniversalUniqueMap();
 
-    void v_InitObject() override;
+    bool ConvergenceCheck(const Array<OneD, const NekDouble> &Residual);
 
 private:
 };