+++ library/LibUtilities/LinearAlgebra/NekLinSysIterCGLoc.cpp
@@ -55,6 +55,8 @@ NekLinSysIterCGLoc::NekLinSysIterCGLoc(
     const NekSysKey &pKey)
     : NekLinSysIter(pSession, vRowComm, nDimen, pKey)
 {
+    m_NekLinSysTolerance = max(m_NekLinSysTolerance, 1.0E-16);
+
     m_isLocal = true;
 }
 
@@ -69,10 +71,8 @@ void NekLinSysIterCGLoc::v_InitObject()
 int NekLinSysIterCGLoc::v_SolveSystem(
     const int nLocal, const Array<OneD, const NekDouble> &pInput,
     Array<OneD, NekDouble> &pOutput, [[maybe_unused]] const int nDir,
-    const NekDouble tol, [[maybe_unused]] const NekDouble factor)
+    [[maybe_unused]] const NekDouble factor)
 {
-    m_tolerance = max(tol, 1.0E-16);
-
     DoConjugateGradient(nLocal, pInput, pOutput);
 
     return m_totalIterations;
@@ -135,12 +135,12 @@ void NekLinSysIterCGLoc::DoConjugateGradient(
     m_totalIterations = 0;
 
     // If input residual is less than tolerance skip solve.
-    if (eps < m_tolerance * m_tolerance * m_rhs_magnitude)
+    if (eps < m_NekLinSysTolerance * m_NekLinSysTolerance * m_rhs_magnitude)
     {
         if (m_verbose && m_root)
         {
             cout << "CG iterations made = " << m_totalIterations
-                 << " using tolerance of " << m_tolerance
+                 << " using tolerance of " << m_NekLinSysTolerance
                  << " (error = " << sqrt(eps / m_rhs_magnitude)
                  << ", rhs_mag = " << sqrt(m_rhs_magnitude) << ")" << endl;
         }
@@ -164,12 +164,12 @@ void NekLinSysIterCGLoc::DoConjugateGradient(
     // Continue until convergence
     while (true)
     {
-        if (m_totalIterations > m_maxiter)
+        if (m_totalIterations > m_NekLinSysMaxIterations)
         {
             if (m_root)
             {
                 cout << "CG iterations made = " << m_totalIterations
-                     << " using tolerance of " << m_tolerance
+                     << " using tolerance of " << m_NekLinSysTolerance
                      << " (error = " << sqrt(eps / m_rhs_magnitude)
                      << ", rhs_mag = " << sqrt(m_rhs_magnitude) << ")" << endl;
             }
@@ -213,12 +213,12 @@ void NekLinSysIterCGLoc::DoConjugateGradient(
         m_totalIterations++;
 
         // Test if norm is within tolerance
-        if (eps < m_tolerance * m_tolerance * m_rhs_magnitude)
+        if (eps < m_NekLinSysTolerance * m_NekLinSysTolerance * m_rhs_magnitude)
         {
             if (m_verbose && m_root)
             {
                 cout << "CG iterations made = " << m_totalIterations
-                     << " using tolerance of " << m_tolerance
+                     << " using tolerance of " << m_NekLinSysTolerance
                      << " (error = " << sqrt(eps / m_rhs_magnitude)
                      << ", rhs_mag = " << sqrt(m_rhs_magnitude) << ")" << endl;
             }