+++ solvers/CompressibleFlowSolver/EquationSystems/CompressibleFlowSystemImplicit.h
@@ -37,6 +37,9 @@
 #define NEKTAR_SOLVERS_COMPRESSIBLEFLOWSOLVER_COMPRESSIBLEFLOWSYSTEMIMPLICIT_H
 
 #include <CompressibleFlowSolver/EquationSystems/CompressibleFlowSystem.h>
+#include <CompressibleFlowSolver/Preconditioner/PreconCfs.h>
+#include <CompressibleFlowSolver/Preconditioner/PreconCfsOp.h>
+#include <LibUtilities/LinearAlgebra/NekNonlinSysIter.h>
 
 namespace Nektar
 {
@@ -74,12 +77,11 @@ protected:
     NekDouble m_TimeIntegLambda = 0.0;
     NekDouble m_inArrayNorm     = -1.0;
     NekDouble m_jacobiFreeEps;
-    NekDouble m_newtonRelativeIteTol;
 
     TensorOfArray4D<NekSingle> m_stdSMatDataDBB;
     TensorOfArray5D<NekSingle> m_stdSMatDataDBDB;
 
-    LibUtilities::NekNonlinSysSharedPtr m_nonlinsol;
+    LibUtilities::NekNonlinSysIterSharedPtr m_nonlinsol;
 
     PreconCfsSharedPtr m_preconCfs;